Added option to pay rent from ATM, fixed the servicePaid bug fixed typo
The servicePaid value never got reset, so as soon as a total of 2000 worth of services were done, rent became free. 
Now when rent is paid servicePaid is set to 0 again.

If Sveta doesn't have cash, but has the money in the ATM she doesn't get evicted, instead she has the option to go down to the ATM and pay in cash.

Fixed the "non-existant window curtain" typo.