[fixed] cleaned up some code that had a lot of copypasting:
(1) made room number a variable passed to the room sublocation instead of having a separate near identical sublocation for each room (this, along with changing the line telling the player what interview room to go to so that it uses the <<room>> variable, also makes it so that only one line needs to be updated if changing what room the SMTV scene takes place in instead of three lines).

(2) split the SMTV scenes out into gt calls as there were two identical copies of each of the scenes in the code previously.

(1) and (2) together reduce the size of this file by 180 lines, or 27%